Consistant scrolling direction
Scrolling up to change a value is semantically associated with
increasing that value.
Volume already worked this way, and the same is true for seeking
by scrolling outside of the UI, but scrolling up on the timeline
seeked backwards and scrolling up on the speed widget reduced speed.
Now scrolling on the timeline or outside of the UI seeks in the same
direction and speed also increases when scrolling up on the widget.