Allow creating empty folders
So far it was not possible to create an empty folder that was then was
later filled up with commands (or folders).
Because of that the structure of input.conf had to be based around
the menu.
By creating empty folders with the `ignore` command, that requirement
has been loosened a bit.
Trimming whitespace from `command` was necessary to check for "ignore".