fix: title hiding when mouse is at `0,0` coordinate
This was done because when player start on linux, the cursor position is reported as `0,0`, even though there is no cursor there, which lead to title being always visible on startup until cursor moves, which was annoying. Previous solution just hardcoded `0,0` coordinate to be recognized as `infinity,infinity`. This one Adds a filter for real and initial cursor position events, and only forces `infinity,infinity` when we haven't received a single real mouse movement yet.

closes #130