fix: tweak timeline progress and chapter elements pixel snapping (#137)
* Progress indicator position not rounded

Having the progress indicator only move by whole pixels leads to juttery
movement. Not rounding the x coordinate, like it used to be before
814edb52cb741e3660aabab18b4497c75d0455aa looks much smoother.

* Round x coordinate of chapter line

Not rounding the x coordinate of the chapter line leads to blurry edges.
This is particularily noticable with thin chapter lines (e.g. 1 px).
Rounding to the nearest integer results in crisp lines.