feat: support `osc-margins` shared-script-property
The documentation cautions against using shared script properties, but
console.lua relies on the margin to not overlap with the bottom bar from
the default osc. Without support for that in uosc, using the console
was always annoying.

Now it also doesn't overlap with the uosc timeline.
The script property gets updated on render whenever at least one of of
the margins changed.