fix: timeline not using real time width
Chapters and cache used the time width calculated from the time padding,
which was a rounded value, instead of the time width that is used for
the progress indicator and the time calculation for mouse input.

This led to e.g. the chapter line for a chapter at 0s being rendered one
pixel too far to the right.

Now the time_x and time_width are the same as in the other calculations
and the chapter and cache components decide for themselves how to deal
with that.

Chapter indicators are now correctly positioned, even at the extremes.

Cache now starts and ends exactly where the center pixel for the
progress line would be. This is inclusive, so e.g. for a 3 pixel wide
progress line, time would start 1.5 pixels in, but cache starts 1 pixel
in. This is to have clean edges for the cache at the start and the end.

The padding indicator for the cache is now moved one pixel outside of
the cache range instead of overlapping with the beginning and the end,
basically making the cache go from one wall to the other.