feat: better buffering detection
Leverage not just information from the cache, but also the player
playback state to get a better idea of when playback is stalled due to
buffering.

Use different metric for enabling the indicator then for disabling it.
This allows the indicator to be continuously on instead of switching
on/off multiple times within the same stall.