fix: remove redundant symbol order from sorting algorithm (#608)
We were already using the same sorting algorithm as autoload.lua,
but with the addition of a custom symbol order. However that symbol
order was only applied to the first character, so when sorting absolute
paths it doesn't actually make a difference.

Using the exact same algorithm for autoload.lua and uosc also makes it
easier to apply any future improvements from one to the other.