fix: don't select not selectable items
It was possible for not selectable items to be selected via page up/down
as well as when updating the menu.

Page down now goes to the next best selectable item after adding to the
index, like is the case for next(), with a fallback to an earlier index.
Page up is the same as page down but inverted direction.

Updating the menu now chooses the closes selectable item to the intended
index.