feat: broadcast uosc version as early as possible
Scripts that offer a UI variant based on the uosc menu might get called
right after startup before uosc manages to respond to their
`get-version` request. As a result the non uosc UI will be shown.
If we proactively broadcast the version as soon as possible then we
minimize the chance of that happening.