feat: added `update` command to update uosc (#700)
`update` command spawns a subprocess using `mp.command_native_async()` that fetches an appropriate installer from uosc's main branch and runs it in powershell or bash.

This is pretty fragile and doesn't work everywhere, such as snap packages on Linux, or mpv on MacOS.

In snap on Linux, I think we are not allowed to use commands like `curl` or `unzip`, even if they're installed. I guess snap's sandboxing just hides them from us, even if we use a direct path to the binary on the system.

On MacOS, we get a weird `(23) Failed writing body` error... no idea.