feat: adjusted menu modifier behavior and added buttons where appropriate
Activating item without closing the menu is now on `alt`. This is just a non-enforced guideline for all menus to implement when appropriate.

Adding files/folders to playlist is now on `shift`, and doesn't close the menu by default.

Actions that change something that is not immediately obvious (like adding files to playlist) now trigger an OSD message with what happened, so that there's some indication that it did in fact happen. I might move this into a custom notifications UI in future.

closes #909