refactor: cursor events in a failed attempt to fix touch
Seems like down events fire BEFORE `mouse-pos` and `touch-pos` are triggered, which means our down event handlers can't tell where the cursor actually is to properly triage the event.

The only fix I see is moving everything to up events, which is not an option since people always complain about unresponsive UI when we do that, so I'm giving up on this for now.

Still, this has some code cleanups and useful changes, so it's at least a refactor instead of a fix. And touch is _somewhat_ usable now (if you tap enough times on something, you get a click!).

ref #1047