fix: `mbtn_left` being forwarded when it shouldn't
We were taking over button events any time there is an interactive element on screen. This didn't seem like an issue because we forward events that we don't handle. Unfortunately that breaks during dragging, when user drags the window and our forwarding triggers `mbtn_left` binding, which wouldn't happen otherwise.

We were enabling event handlers when there's any interactive element rendered to make touch a bit more usable, but it was still pretty broken, so I'm reverting that and enabling handlers only when cursor is above an interactive element.

ref #1118