feat(api): added `Menu.bind_keys` option
When menu was open, it was taking over too many keys that people had keybinds for. This didn't seem like a problem because they were only active while the menu was open, but there were complains anyway. Since this commit, if you want to listen for non-common keys that are not bound by default like `f1-f12`, you need to tell the menu to bind the listeners for them with `bind_keys` option. It accepts a table with keybinds such as `{'f1', 'alt+f2'}` etc.

closes #996, #1127